`internal/controllers/**`: Refactor
`internal/controllers/kube_evoy_config_manager.go`
-------------------------------------------------

Renamed from `internal/controllers/config_manager.go`.

`internal/controllers/parser.go`
--------------------------------

`parser.go` is too large so move the non-member functions to different files.
